+++ CHANGELOG.rst
@@ -23,6 +23,7 @@ Changed
 Added
 ~~~~~
 - Testing on MacOS.
+- :option:`--exe-test-module` cli option for setting the default test module
 
 0.4.0 - 2020-05-03
 ------------------